Brickly: Alternate asyncio usage fix for Python 3.7+ that is compatible with Python 3.4.2
TX-Pi is still based on Raspbian Jessie, which comes with Python 3.4.2, which is not only incompatible
with the new async/await syntax but does not even have the asyncio.ensure_future alternative to
asyncio.async.

In order to stay compatible with all existing Python versions >= 3.4.2, resort to this crude getattr
hack.

Note: This will crash and burn if neither asyncio.async not asyncio.ensure_future are present.